+++ src/typeorm/entities/announcement/AnnouncementConfig.ts
@@ -1,0 +1,22 @@
+import { Column, CreateDateColumn, Entity, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm';
+
+@Entity()
+export class AnnouncementConfig {
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column()
+    guildId: string;
+
+    @Column()
+    minecraftRoleId: string;
+
+    @Column()
+    defaultChannelId: string;
+
+    @CreateDateColumn()
+    createdAt: Date;
+
+    @UpdateDateColumn()
+    updatedAt: Date;
+}